refactor: loadSample into lib module
