+++ wp-content/plugins/audioblocks/src/frontend.js
@@ -1,19 +1,8 @@
 
+import loadSample from './lib/preload-audio-sample';
 
 const AudioContext = window.AudioContext || window.webkitAudioContext;
 const audioContext = new AudioContext();
-
-function loadSample(url, audioContext, callback) {
-	var request = new XMLHttpRequest();
-	request.open('GET', url);
-	request.responseType = 'arraybuffer';
-	request.onload = function () {
-		console.log('sample loaded, decoding');
-		audioContext.decodeAudioData(request.response, callback);
-	};
-	console.log('loading sample');
-	request.send();
-}
 
 function getSecondsPerBeat( tempoBpm ) {
 	return 60.0 / tempoBpm;
